Rename petals.{bloom,llama} -> petals.models.{bloom,llama}
