--- src/petals/bloom/__init__.py
@@ -1,7 +1,0 @@
-from petals.bloom.block import WrappedBloomBlock
-from petals.bloom.config import DistributedBloomConfig
-from petals.bloom.model import (
-    DistributedBloomForCausalLM,
-    DistributedBloomForSequenceClassification,
-    DistributedBloomModel,
-)
