--- src/petals/llama/__init__.py
@@ -1,2 +1,0 @@
-from petals.llama.block import WrappedLlamaBlock
-from petals.llama.config import DistributedLlamaConfig
