+++ src/petals/models/__init__.py
@@ -1,0 +1,2 @@
+from petals.models.bloom import *
+from petals.models.llama import *
