+++ src/petals/models/llama/__init__.py
@@ -1,0 +1,2 @@
+from petals.models.llama.block import WrappedLlamaBlock
+from petals.models.llama.config import DistributedLlamaConfig
