+++ src/petals/models/llama/config.py
@@ -1,22 +1,22 @@
 import os
 from typing import Optional, Union
 
-from transformers.models.bloom import BloomConfig
-from transformers.models.bloom.modeling_bloom import BloomAttention
+from transformers.models.llama import LlamaConfig
+from transformers.models.llama.modeling_llama import LlamaAttention
 
-from petals.bloom.block import WrappedBloomBlock
 from petals.client.modeling_utils import LMHeadConfig
 from petals.client.routing.sequence_manager import SequenceManagerConfig
+from petals.models.llama.block import WrappedLlamaBlock
 from petals.utils.auto_config import AutoDistributedConfig
 
 
-class DistributedBloomConfig(BloomConfig, SequenceManagerConfig, LMHeadConfig):
+class DistributedLlamaConfig(LlamaConfig, SequenceManagerConfig, LMHeadConfig):
     pre_seq_len: int = 0  # a number of tokens for prompt tuning.
     tuning_mode: Optional[str] = None  # fine-tuning regime, one of [None, "ptune", "deep_ptune"]
 
-    block_class = WrappedBloomBlock
-    attn_class = BloomAttention
-    block_prefix = "h"
+    block_class = WrappedLlamaBlock
+    attn_class = LlamaAttention
+    block_prefix = "model.layers"
 
     @classmethod
     def from_pretrained(
@@ -27,4 +27,4 @@ class DistributedBloomConfig(BloomConfig, SequenceManagerConfig, LMHeadConfig):
         return super().from_pretrained(model_name_or_path, *args, dht_prefix=dht_prefix, **kwargs)
 
 
-AutoDistributedConfig.register(DistributedBloomConfig)
+AutoDistributedConfig.register(DistributedLlamaConfig)
